Catch all possible migration errors

+++ database/migrations/2016_12_22_150431_changes_for_v430.php
@@ -46,26 +46,28 @@ class ChangesForV430 extends Migration
      */
     public function up(): void
     {
-        try {
-            Schema::create(
-                'available_budgets',
-                static function (Blueprint $table) {
-                    $table->increments('id');
-                    $table->timestamps();
-                    $table->softDeletes();
-                    $table->integer('user_id', false, true);
-                    $table->integer('transaction_currency_id', false, true);
-                    $table->decimal('amount', 32, 12);
-                    $table->date('start_date');
-                    $table->date('end_date');
+        if (!Schema::hasTable('available_budgets')) {
+            try {
+                Schema::create(
+                    'available_budgets',
+                    static function (Blueprint $table) {
+                        $table->increments('id');
+                        $table->timestamps();
+                        $table->softDeletes();
+                        $table->integer('user_id', false, true);
+                        $table->integer('transaction_currency_id', false, true);
+                        $table->decimal('amount', 32, 12);
+                        $table->date('start_date');
+                        $table->date('end_date');
 
-                    $table->foreign('transaction_currency_id')->references('id')->on('transaction_currencies')->onDelete('cascade');
-                    $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-                }
-            );
-        } catch (QueryException $e) {
-            Log::error(sprintf('Could not create table "available_budgets": %s', $e->getMessage()));
-            Log::error('If this table exists already (see the error message), this is not a problem. Other errors? Please open a discussion on GitHub.');
+                        $table->foreign('transaction_currency_id')->references('id')->on('transaction_currencies')->onDelete('cascade');
+                        $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                    }
+                );
+            } catch (QueryException $e) {
+                Log::error(sprintf('Could not create table "available_budgets": %s', $e->getMessage()));
+                Log::error('If this table exists already (see the error message), this is not a problem. Other errors? Please open a discussion on GitHub.');
+            }
         }
     }
 }
