+++ database/migrations/2018_11_06_172532_changes_for_v479.php
@@ -41,16 +41,18 @@ class ChangesForV479 extends Migration
      */
     public function down(): void
     {
-        try {
-            Schema::table(
-                'transaction_currencies',
-                static function (Blueprint $table) {
-                    $table->dropColumn(['enabled']);
-                }
-            );
-        } catch (QueryException|ColumnDoesNotExist $e) {
-            Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
-            Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+        if(Schema::hasColumn('transaction_currencies', 'enabled')) {
+            try {
+                Schema::table(
+                    'transaction_currencies',
+                    static function (Blueprint $table) {
+                        $table->dropColumn(['enabled']);
+                    }
+                );
+            } catch (QueryException|ColumnDoesNotExist $e) {
+                Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
+                Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+            }
         }
     }
 
@@ -61,16 +63,18 @@ class ChangesForV479 extends Migration
      */
     public function up(): void
     {
-        try {
-            Schema::table(
-                'transaction_currencies',
-                static function (Blueprint $table) {
-                    $table->boolean('enabled')->default(0)->after('deleted_at');
-                }
-            );
-        } catch (QueryException $e) {
-            Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
-            Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+        if(!Schema::hasColumn('transaction_currencies', 'enabled')) {
+            try {
+                Schema::table(
+                    'transaction_currencies',
+                    static function (Blueprint $table) {
+                        $table->boolean('enabled')->default(0)->after('deleted_at');
+                    }
+                );
+            } catch (QueryException $e) {
+                Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
+                Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+            }
         }
     }
 }
